+++ src/redi/cli/prompt_util.py
@@ -1,32 +1,9 @@
 from prompt_toolkit import Application, prompt
 from prompt_toolkit.key_binding import KeyBindings
-from prompt_toolkit.keys import Keys
 from prompt_toolkit.layout import HSplit, Layout, Window
 from prompt_toolkit.layout.controls import FormattedTextControl
 
 from redi.i18n import messages
-
-
-def digit_only_key_bindings() -> KeyBindings:
-    kb = KeyBindings()
-
-    @kb.add(Keys.Any)
-    def _(event) -> None:
-        if event.data.isdigit():
-            event.current_buffer.insert_text(event.data)
-
-    return kb
-
-
-def digit_and_period_key_bindings() -> KeyBindings:
-    kb = KeyBindings()
-
-    @kb.add(Keys.Any)
-    def _(event) -> None:
-        if event.data.isdigit() or event.data == ".":
-            event.current_buffer.insert_text(event.data)
-
-    return kb
 
 
 def confirm_delete(summary: str) -> None:
